// compiler/src/diagnostics/generated.ts
/*
 * This file was generated by a script. Do not modify.
 */

import { DiagnosticEntryTable } from './types';

export enum DiagnosticCode {
  INFINITE_STRING = 101,
  USE_DOUBLE_QUOTES = 102,
  UNEXPECTED_SYMBOL = 103,
  DUMMY_ERROR = 0,
}

export const diagnosticCodeTable: DiagnosticEntryTable = {
  101: {
    title: 'Infinite string',
    preamble:
      'I reached the end of this line without seeing an appropriate close quote:',
    message:
      'Strings should be surrounded with double quotes like "this". Please add a closing double quote in your code.',
    code: 101,
  },
  102: {
    title: 'Use double quotes',
    preamble:
      'The following string uses single quotes when it should use double quotes:',
    message:
      "In Arizona, double quotes are used for strings like \"foobar\", whereas it uses single quotes for individual characters like 'a' or '花'.",
    code: 102,
  },
  103: {
    title: 'Unexpected symbol',
    preamble: 'I could not understand the following symbol:',
    code: 103,
  },
  0: {
    title: 'Dummy error',
    preamble:
      'This error is just a placeholder and should not actually be used in code.',
    message:
      'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nam tempor, leo quis dapibus varius, sapien nunc pellentesque ex, vel sollicitudin massa velit nec quam. Nunc neque magna, dictum et tellus imperdiet, tempor dignissim risus. Sed tempus convallis ipsum, non lacinia nisi volutpat vel.',
    code: 0,
  },
};